fix: adjusted navigation group actions and queries
