+++ apps/tailwind-components/app/utils/cms.ts
@@ -15,6 +15,8 @@ import type {
   ICmsOrderWithBlockId,
 } from "../../types/CmsComponents";
 
+import { AddNavigationCard, AddNavigationGroup } from "./cms/add";
+
 export function randomId(): string {
   return crypto.randomUUID();
 }
@@ -399,31 +401,28 @@ export async function addComponent(
   componentType: string
 ) {
   await prepareOrder(schema, order, parentBlock);
-  let insertedComponent = false;
-  switch (componentType) {
-    case "Paragraph":
-      await AddParagraph(schema, id);
-      insertedComponent = true;
-      break;
-    case "Heading":
-      await AddHeading(schema, id);
-      insertedComponent = true;
-      break;
-    case "Image":
-      await AddImage(schema, id);
-      insertedComponent = true;
-      break;
-    case "NavigationGroups":
-      await AddNavigationGroup(schema, id);
-      break;
-    case "NavigationCards":
-      await AddNavigationCard(schema, id, parentBlock);
-      break;
+
+  if (componentType === "Paragraph") {
+    await AddParagraph(schema, id);
   }
 
-  if (insertedComponent) {
-    await AddOrder(schema, id, order, parentBlock);
+  if (componentType === "Heading") {
+    await AddHeading(schema, id);
   }
+
+  if (componentType === "Image") {
+    await AddImage(schema, id);
+  }
+
+  if (componentType === "NavigationGroups") {
+    await AddNavigationGroup(schema, id);
+  }
+
+  if (componentType === "NavigationCards") {
+    await AddNavigationCard(schema, id, parentBlock);
+  }
+
+  await AddOrder(schema, id, order, parentBlock);
 }
 
 export async function addBlock(
@@ -466,39 +465,6 @@ async function AddHeader(schema: string, id: string) {
         id: `${id}`,
         title: "Title",
         subtitle: "A subtitle here",
-      },
-    ],
-  };
-
-  await cmsFetch(schema, query, variables);
-}
-
-async function AddNavigationGroup(schema: string, id: string) {
-  const query = `mutation insert($nav:[NavigationGroupsInput]) {
-    insert(NavigationGroups: $nav) {
-      status
-      message
-    }
-  }`;
-  const variables = { nav: [{ id: id }] };
-  await cmsFetch(schema, query, variables);
-}
-
-async function AddNavigationCard(schema: string, id: string, parentId: string) {
-  const query = `mutation insert($element: [NavigationCardsInput]) {
-    insert(NavigationCards: $element) {
-      status
-      message
-    }
-  }`;
-  const variables = {
-    element: [
-      {
-        id: id,
-        title: "Title",
-        description: "A description about the link",
-        url: "https://molgenis.org",
-        displayedInNavigationGroup: { id: parentId },
       },
     ],
   };
